Fix loader function
